feat: extrating components part 2
