+++ lib/live_view_phoenix_web/live/volunteers_live.ex
@@ -56,27 +56,6 @@ defmodule LiveViewPhoenixWeb.VolunteersLive do
     {:noreply, assign(socket, changeset: changeset)}
   end
 
-  def handle_event("toggle-states", %{"id" => id}, socket) do
-    volunteer = Volunteers.get_volunteer!(id)
-
-    {:ok, _volunteer} =
-      Volunteers.update_volunteer(
-        volunteer,
-        %{checked_out: !volunteer.checked_out}
-      )
-
-    # volunteers = Volunteers.list_volunteers()
-
-    # socket =
-    #   assign(socket,
-    #   volunteers: volunteers
-    #   )
-
-    # :timer.sleep(500)
-
-    {:noreply, socket}
-  end
-
   def handle_info({:volunteer_created, volunteer}, socket) do
     socket =
       update(
