+++ lib/live_view_phoenix_web/live/volunteers_live.html.leex
@@ -13,20 +13,7 @@
 </form>
   <div >
     <%= for volunteer <- @volunteers do %>
-      <div class="volunteer <%= if volunteer.checked_out, do: "out" %>" id="<%= volunteer.id%>">
-        <div class="name">
-          <%= volunteer.name %>
-        </div>
-        <div class="phone">
-          <img src="images/phone.svg">
-          <%= volunteer.phone %>
-        </div>
-        <div class="status">
-          <button phx-click="toggle-states" phx-value-id="<%= volunteer.id %>" phx_disable_with="Saving.." class="<%= if volunteer.checked_out, do: "out" %>">
-            <%= if volunteer.checked_out, do: "Check In", else: "Check Out" %>
-          </button>
-        </div>
-      </div>
+      <%= live_component @socket, LiveViewPhoenixWeb.VolunnterComponent, volunteer: volunteer, id: volunteer.id %>
     <% end %>
   </div>
 </div>
